+++ packages/sources/svelte/rollup.config.js
@@ -19,8 +19,18 @@ const src = {
       components: ['VtmnBadge', 'VtmnPrice', 'VtmnRating', 'VtmnTag'],
     },
     {
+      folder: 'navigation',
+      components: ['VtmnSearch'],
+    },
+    {
       folder: 'overlays',
-      components: ['VtmnAlert', 'VtmnModal', 'VtmnPopover', 'VtmnSnackbar'],
+      components: [
+        'VtmnAlert',
+        'VtmnModal',
+        'VtmnPopover',
+        'VtmnSnackbar',
+        'VtmnToast',
+      ],
     },
     {
       folder: 'selection-controls',
@@ -34,6 +44,7 @@ const src = {
         'VtmnDivider',
         'VtmnList',
         'VtmnListItem',
+        'VtmnSkeleton',
       ],
     },
   ],